feat: update Python version support and enhance ISyntax image wrapper
