+++ README.md
@@ -8,7 +8,7 @@
 
 <div align="center">
 
-[![Python](https://img.shields.io/badge/Python-3.8+-blue.svg)](https://python.org)
+[![Python](https://img.shields.io/badge/Python-3.10+-blue.svg)](https://python.org)
 [![License](https://img.shields.io/badge/License-MIT-green.svg)](LICENSE)
 [![PyPI](https://img.shields.io/badge/PyPI-tissuelab-blue.svg)](https://pypi.org/project/tissuelab/)
 
@@ -45,7 +45,7 @@ with DicomImageWrapper("path/to/image.dcm") as dicom:
 - **DICOM**: Medical imaging standard
 - **NIfTI**: Neuroimaging format
 - **CZI**: Zeiss microscopy format (Windows)
-- **ISyntax**: Philips pathology format (Windows)
+- **ISyntax**: Philips pathology format
 - **Simple Images**: JPEG, PNG, BMP, etc.
 
 ## 🔧 API Reference
@@ -59,7 +59,7 @@ from tissuelab_sdk.wrapper import (
     NiftiImageWrapper,   # NIfTI files
     SimpleImageWrapper,  # JPEG, PNG, etc.
     CziImageWrapper,     # CZI files (Windows)
-    ISyntaxImageWrapper  # ISyntax files (Windows)
+    ISyntaxImageWrapper  # ISyntax files
 )
 
 # All wrappers share the same interface
