+++ requirements.txt
@@ -1,9 +1,10 @@
-pillow~=11.0.0
-tifffile>=2023.0.0
-tiffslide
-pydicom>=2.3.0
-numpy==1.26.4
-nibabel>=5.2.0
+pillow>=11.0.0
+tifffile>=2026.3.3
+tiffslide==3.0.0
+pydicom>=3.0.2
+numpy>=1.26.4,<2.0
+nibabel==5.4.2
+pyisyntax==0.1.5
 czifile>=2019.7.2; platform_system == "Windows"
 pylibCZIrw>=3.3.0; platform_system == "Windows"
 pywin32; platform_system == "Windows"
