+++ tissuelab_sdk/__init__.py
@@ -3,6 +3,7 @@
 This package provides the public import path `tissuelab_sdk.wrapper`.
 """
 
+__version__ = "0.1.14"
 __all__ = ["wrapper"]
