+++ tissuelab_sdk/wrapper/windows.py
@@ -1,16 +1,13 @@
 from .common import *  # re-export shared wrappers
 from PIL import Image
 import numpy as np
+import io
 from pylibCZIrw import czi
 import threading
 import time
 import pythoncom
 import czifile
 import xml.etree.ElementTree as ET
-try:
-    from isyntax import ISyntax
-except Exception:  # pragma: no cover
-    ISyntax = None
 class CziImageWrapper:
 
     def __init__(self, czi_path, max_levels=5):
@@ -111,75 +108,3 @@ class CziImageWrapper:
             pil_img = Image.fromarray(img)
             return np.array(pil_img) if as_array else pil_img
 
-
-class ISyntaxImageWrapper:
-    """Wrapper for ISyntax image files using WSI interface"""
-
-    def __init__(self, isyntax_path):
-        self.path = isyntax_path
-        self.isyntax_reader = None
-        try:
-            self.isyntax_reader = ISyntax.open(self.path)
-            self._init_levels()
-        except Exception as e:
-            print(f"Error opening ISyntax file {isyntax_path}: {e}")
-            # Set default values if initialization fails
-            self.dimensions = (100, 100)
-            self.level_count = 1
-            self.level_dimensions = [(100, 100)]
-            self.properties = {
-                'vendor': 'ISyntaxImageWrapper',
-                'dimensions': '100x100',
-                'level_count': '1',
-                'error': str(e)
-            }
-            raise
-    
-    def close(self):
-        """Explicitly close the ISyntax reader."""
-        if hasattr(self, 'isyntax_reader') and self.isyntax_reader is not None:
-            try:
-                self.isyntax_reader.close()
-            except Exception as e:
-                print(f"Error closing ISyntax reader: {e}")
-            finally:
-                self.isyntax_reader = None
-    
-    def __del__(self):
-        """Safe destructor that handles missing attributes."""
-        try:
-            if hasattr(self, 'isyntax_reader') and self.isyntax_reader is not None:
-                self.close()
-        except Exception as e:
-            # Suppress errors during cleanup
-            pass
-
-    def _init_levels(self):
-        if self.isyntax_reader is None:
-            raise RuntimeError("ISyntax reader not initialized")
-        
-        self.dimensions = self.isyntax_reader.dimensions
-        self.level_count = self.isyntax_reader.level_count
-        self.level_dimensions = self.isyntax_reader.level_dimensions
-
-        self.properties = {
-            'vendor': 'ISyntaxImageWrapper',
-            'dimensions': f'{self.dimensions[0]}x{self.dimensions[1]}',
-            'level_count': str(self.level_count),
-        }
-
-    def read_region(self, location, level, size, as_array=False):
-        """Read a region from the image at the specified level, array type: RGBA"""
-        if self.isyntax_reader is None:
-            raise RuntimeError("ISyntax reader not initialized")
-        
-        # convert location and size to integers
-        scale_factor = self.dimensions[0] / self.level_dimensions[level][0]
-        x, y = location
-        scaled_x = int(x / scale_factor)
-        scaled_y = int(y / scale_factor)
-        scaled_w = int(size[0] / scale_factor)
-        scaled_h = int(size[1] / scale_factor)
-        img_array = self.isyntax_reader.read_region(
-            scaled_x, scaled_y, scaled_w, scaled_h, level)
-        return img_array if as_array else Image.fromarray(img_array, mode='RGBA')
